Fixed imports and renamed invadedhouse directory
